feat(expo-notifications): apply review comments changes
