+++ src/hooks/index.ts
@@ -74,6 +74,7 @@ export * from './useAppStateActive'
 export * from './useAuth'
 export * from './useBoolean'
 export * from './useCachedResources'
+export * from './useNotificationSetup'
 export * from './useSecurePassword'
 export * from './useTimestamp'
 export * from './useToggle'
